created SFcomparison code, copying necessary code from MCMAC (related to issue #5)
This is the base of some code that I am writing to calculate the relative velocity and
time-since-collsion of the Springel and Farrar (2007) Bullet Cluster using my model from
Dawson (2012).  This will be used to test the accuracy of my model with that of their
hydrodynamic simulations.  At this commit I have simply copied necessary code from the
MCMAC program v0.1. In following commits I will modify the code to obtain the necessary
output format for comparison with the Springel and Farrar results.